requirements.txt
Pillow==4.3.0
django==1.11.7
django-crispy-forms==1.6.0